+++ src/main/java/com/boot/quantitymeasurement/service/WeightUnit.java
@@ -8,13 +8,13 @@ public class WeightUnit implements MainUnit {
 
     @Override
     public Quantity getConvertedQuantity(Quantity quantity) {
-        int intoBaseUnit = getIntoBaseUnit(quantity.getUnitOne(), quantity.getSizeOne());
+        double intoBaseUnit = getIntoBaseUnit(quantity.getUnitOne(), quantity.getSizeOne());
         quantity.setSizeTwo(getInConvertedUnit(quantity.getUnitTwo(),intoBaseUnit));
         return quantity;
     }
 
     @Override
-    public int getInConvertedUnit(Unit.SubUnit unitTwo, int intoBaseUnit) {
+    public double getInConvertedUnit(Unit.SubUnit unitTwo, double intoBaseUnit) {
         switch (unitTwo){
             case KG:
                 return intoBaseUnit/1000;
@@ -25,7 +25,7 @@ public class WeightUnit implements MainUnit {
         }
     }
 
-    public int getIntoBaseUnit(Unit.SubUnit unitOne, int sizeOne) {
+    public double getIntoBaseUnit(Unit.SubUnit unitOne, double sizeOne) {
         switch (unitOne){
             case KG:
                 return sizeOne*1000;
